# app/backend/docs/swagger.yaml
basePath: /
definitions:
...
  domain.LoginInput:
    properties:
      email:
        example: marie@example.com
        type: string
      password:
        type: string
    type: object
...
  domain.RefreshInput:
    properties:
      refresh_token:
        type: string
    type: object
  domain.RegisterInput:
    properties:
      display_name:
        maxLength: 80
        type: string
      email:
        example: marie@example.com
        type: string
      locale:
        enum:
        - fr
        - en
        example: fr
        type: string
      password:
        minLength: 8
        type: string
    type: object
...
  domain.UpdateProfileInput:
    properties:
      display_name:
        maxLength: 80
        type: string
      locale:
        enum:
        - fr
        - en
        type: string
    type: object
...
  httpapi.AuthResponse:
    properties:
      data:
        $ref: '#/definitions/domain.AuthResult'
    type: object
...
  httpapi.ContributionsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.Contribution'
        type: array
      meta:
        $ref: '#/definitions/httpapi.PaginationMeta'
    type: object
...
  httpapi.Problem:
    properties:
      detail:
        type: string
      status:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
...
  httpapi.ProductPriceResponse:
    properties:
      data:
        $ref: '#/definitions/domain.ProductPrice'
    type: object
...
  httpapi.UserResponse:
    properties:
      data:
        $ref: '#/definitions/domain.User'
    type: object
info:
  contact: {}
  description: |-
    API de comparaison du coût de la vie dans la zone euro.
    Les montants décimaux sont exposés sous forme de chaînes pour préserver leur précision.
  title: OjoZone API
  version: "1.0"
paths:
...
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Identifiants
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Ouvrir une session
      tags:
      - Authentification
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      parameters:
      - description: Jeton de rafraîchissement
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.RefreshInput'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Révoquer la session
      tags:
      - Authentification
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Un jeton de rafraîchissement ne peut être utilisé qu'une seule
        fois.
      parameters:
      - description: Jeton de rafraîchissement
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.RefreshInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Renouveler les jetons
      tags:
      - Authentification
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Valide l'email, le mot de passe et la langue, puis retourne le
        profil et deux jetons.
      parameters:
      - description: Inscription
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.RegisterInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/httpapi.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Créer un compte et ouvrir une session
      tags:
      - Authentification
...
  /api/v1/me:
    delete:
      description: Révoke les sessions et anonymise les données conservées à des fins
        statistiques.
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.Problem'
      security:
      - BearerAuth: []
      summary: Supprimer le compte
      tags:
      - Compte
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.Problem'
      security:
      - BearerAuth: []
      summary: Lire le profil courant
      tags:
      - Compte
    patch:
      consumes:
      - application/json
      description: 'Les champs absents conservent leur valeur existante ; display_name:
        null efface le nom.'
      parameters:
      - description: Champs à modifier
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateProfileInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.Problem'
      security:
      - BearerAuth: []
      summary: Modifier le profil courant
      tags:
      - Compte
  /api/v1/me/contributions:
    get:
      description: Retourne les prix produits et carburants soumis, quel que soit
        leur statut.
      parameters:
      - default: 20
        description: Nombre de résultats
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: Décalage
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.ContributionsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.Problem'
      security:
      - BearerAuth: []
      summary: Suivre ses contributions
      tags:
      - Compte
...
  /api/v1/product-prices:
    get:
      parameters:
      - description: UUID du produit
        format: uuid
        in: query
        name: product_id
        type: string
      - description: UUID de la zone géographique
        format: uuid
        in: query
        name: geo_area_id
        type: string
      - description: Statut
        enum:
        - pending
        - approved
        - rejected
        - flagged
        in: query
        name: status
        type: string
      - default: 20
        description: Nombre de résultats
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: Décalage
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.ProductPriceListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Lister toutes les observations de prix produit
      tags:
      - Prix
    post:
      consumes:
      - application/json
      description: L'upsert utilise la clé composée source_id et source_record_id.
        source_record_id est donc obligatoire.
      parameters:
      - description: Observation de prix
        in: body
        name: price
        required: true
        schema:
          $ref: '#/definitions/domain.ProductPriceUpsert'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.ProductPriceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Créer ou mettre à jour un prix produit
      tags:
      - Prix
  /api/v1/product-prices/{id}:
    delete:
      parameters:
      - description: UUID de l'observation
        format: uuid
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Supprimer un prix produit
      tags:
      - Prix
    get:
      parameters:
      - description: UUID de l'observation
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.ProductPriceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Obtenir un prix produit
      tags:
      - Prix
    put:
      consumes:
      - application/json
      description: Crée l'observation si l'UUID n'existe pas, sinon remplace ses champs
        modifiables.
      parameters:
      - description: UUID de l'observation
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Nouvel état du prix
        in: body
        name: price
        required: true
        schema:
          $ref: '#/definitions/domain.ProductPriceUpsert'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.ProductPriceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Remplacer un prix produit
      tags:
      - Prix
  /api/v1/products:
    get:
      parameters:
      - description: Nom, marque ou code-barres
        in: query
        name: q
        type: string
      - default: 20
        description: Nombre de résultats
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: Décalage
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.ProductListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Lister les produits
      tags:
      - Produits
    post:
      consumes:
      - application/json
      description: Avec un code-barres, l'upsert utilise le code-barres. Sans code-barres,
        un UUID id est obligatoire.
      parameters:
      - description: Produit à créer ou mettre à jour
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/domain.ProductUpsert'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.ProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Créer ou mettre à jour un produit
      tags:
      - Produits
...
  /api/v1/products/{id}/prices:
    get:
      parameters:
      - description: UUID du produit
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: UUID de la zone géographique
        format: uuid
        in: query
        name: geo_area_id
        type: string
      - default: 20
        description: Nombre de résultats
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: Décalage
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.ProductPriceListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Lister les prix d'un produit
      tags:
      - Prix
  /api/v1/products/by-barcode/{barcode}:
    get:
      parameters:
      - description: Code-barres
        in: path
        maxLength: 32
        name: barcode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.ProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Obtenir un produit par code-barres
      tags:
      - Produits
...
  /ops/readiness:
    get:
      description: Vérifie notamment la connexion PostgreSQL.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.StatusResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Vérifier la disponibilité
      tags:
      - Operations
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"